#24: 機能の分離を完了
