+++ typescript/parser/markdown.ts
@@ -1,12 +1,8 @@
 import * as mdocx from './types';
+import * as mathjax from './mathjax';
+import * as line from './line';
 import { z } from 'zod';
 import YAML from 'yaml';
-import { mathjax } from 'mathjax-full/js/mathjax';
-import { TeX } from 'mathjax-full/js/input/tex';
-import { SVG } from 'mathjax-full/js/output/svg';
-import { AllPackages } from 'mathjax-full/js/input/tex/AllPackages';
-import { liteAdaptor } from 'mathjax-full/js/adaptors/liteAdaptor';
-import { RegisterHTMLHandler } from 'mathjax-full/js/handlers/html';
 
 const zodMarkdownHeader = z.object({
   title: z.string().default(''),
@@ -14,89 +10,14 @@ const zodMarkdownHeader = z.object({
   author: z.string(),
 });
 
-function mathjax2svg(expression: string): string {
-  const adaptor = liteAdaptor();
-  RegisterHTMLHandler(adaptor);
-
-  const document = mathjax.document('', {
-    InputJax: new TeX({ packages: AllPackages }),
-    OutputJax: new SVG({ fontCache: 'none' }),
-  });
-
-  return adaptor.innerHTML(document.convert(expression));
-}
-
-function parseText(text: string): mdocx.MdocxText[] {
-  const reSpliter = /(\*\*.+\*\*|\*.+\*|\$.+\$|\[\^.+\])/;
-  const reBold = /^\*\*(.+)\*\*$/;
-  const reItalic = /^\*(.+)\*$/;
-  const reMathjax = /^\$(.+)\$$/;
-  const reReference = /^\[\^(.+)\]$/;
-
-  function parseTextInternal(
-    text: string,
-    options: mdocx.MdocxTextOptions
-  ): mdocx.MdocxText[] {
-    const splitedTexts = text.split(reSpliter);
-    const result: mdocx.MdocxText[] = [];
-
-    for (const splitedText of splitedTexts) {
-      if (splitedText === undefined || splitedText === null) {
-        continue;
-      }
-
-      const boldMatched = splitedText.match(reBold);
-      if (boldMatched !== null) {
-        result.push(
-          ...parseTextInternal(boldMatched[1], { ...options, bold: true })
-        );
-        continue;
-      }
-
-      const italicMatched = splitedText.match(reItalic);
-      if (italicMatched !== null) {
-        result.push(
-          ...parseTextInternal(italicMatched[1], { ...options, italic: true })
-        );
-        continue;
-      }
-
-      const mathjaxMatched = splitedText.match(reMathjax);
-      if (mathjaxMatched !== null) {
-        result.push({
-          type: 'svg',
-          svg: mathjax2svg(mathjaxMatched[1]),
-        });
-        continue;
-      }
-
-      if (splitedText === null) {
-        throw new Error('splitedText is null');
-      }
-      const referenceMatched = splitedText.match(reReference);
-      if (referenceMatched !== null) {
-        result.push({
-          type: 'reference',
-          key: referenceMatched[1],
-        });
-        continue;
-      }
-
-      result.push({
-        type: 'normal',
-        expression: splitedText,
-        options: options,
-      });
-    }
-
-    return result;
-  }
-
-  return parseTextInternal(text, {
-    bold: false,
-    italic: false,
-  });
-}
+const reEmpty = /^\s*$/;
+const reSeparator = /^---\s*$/;
+const reHeader = /^(#{1,6})\s+/;
+const reImage = /^!\[(.*)\]\((.+)\)$/;
+const reImageDescription = /^>\s/;
+const reReference = /^\[\^(.+)\]:\s+(.+)$/;
+const reMathjaxLq = /^\$\$/;
+const reMathjaxRq = /\$\$\s*$/;
 
 function ParseMarkdown(
   markdown: string,
@@ -104,15 +25,6 @@ function ParseMarkdown(
   destination: string,
   styleTemplate?: string
 ): mdocx.MdocxDocument {
-  const reEmpty = /^\s*$/;
-  const reSeparator = /^---\s*$/;
-  const reHeader = /^(#{1,6})\s+/;
-  const reImage = /^!\[(.*)\]\((.+)\)$/;
-  const reImageDescription = /^>\s/;
-  const reReference = /^\[\^(.+)\]:\s+(.+)$/;
-  const reMathjaxLq = /^\$\$/;
-  const reMathjaxRq = /\$\$\s*$/;
-
   const mdlines = markdown.split(/\r\n|\n/);
   let lineIdx = 0;
   const header = (function () {
@@ -181,7 +93,7 @@ function ParseMarkdown(
           content: {
             type: 'heading',
             level: level,
-            text: parseText(mdlines[lineIdx].replace(reHeader, '')),
+            text: line.parseText(mdlines[lineIdx].replace(reHeader, '')),
           },
         });
       }
@@ -226,7 +138,7 @@ function ParseMarkdown(
         content: {
           type: 'image',
           filenames: filenames,
-          description: parseText(description),
+          description: line.parseText(description),
         },
       });
 
@@ -243,7 +155,7 @@ function ParseMarkdown(
         status: status,
         key: key,
         displayName: '',
-        description: parseText(description),
+        description: line.parseText(description),
       };
       continue;
     }
@@ -270,7 +182,7 @@ function ParseMarkdown(
         status: status,
         content: {
           type: 'mathjax',
-          svg: mathjax2svg(expression),
+          svg: mathjax.mathjax2svg(expression),
         },
       });
 
@@ -281,7 +193,7 @@ function ParseMarkdown(
       status: status,
       content: {
         type: 'normal',
-        text: parseText(mdlines[lineIdx]),
+        text: line.parseText(mdlines[lineIdx]),
       },
     });
   }
