+++ typescript/parser/mathjax.ts
@@ -1,0 +1,20 @@
+import { mathjax } from 'mathjax-full/js/mathjax';
+import { TeX } from 'mathjax-full/js/input/tex';
+import { SVG } from 'mathjax-full/js/output/svg';
+import { AllPackages } from 'mathjax-full/js/input/tex/AllPackages';
+import { liteAdaptor } from 'mathjax-full/js/adaptors/liteAdaptor';
+import { RegisterHTMLHandler } from 'mathjax-full/js/handlers/html';
+
+function mathjax2svg(expression: string): string {
+  const adaptor = liteAdaptor();
+  RegisterHTMLHandler(adaptor);
+
+  const document = mathjax.document('', {
+    InputJax: new TeX({ packages: AllPackages }),
+    OutputJax: new SVG({ fontCache: 'none' }),
+  });
+
+  return adaptor.innerHTML(document.convert(expression));
+}
+
+export { mathjax2svg };
